refactor: chat 타입 상수화 및 코딩 컨벤션 적용
