+++ src/pages/ChatRoomListPage/ChatRoomListPage.style.ts
@@ -21,8 +21,6 @@ export const Main = styled.div`
   margin: 10px 0;
 `;
 
-export const TabBar = styled(Flex)``;
-
 export const TabBarButton = styled.button<{ isSelected: boolean }>`
   width: 50%;
   background-color: white;
@@ -64,8 +62,6 @@ export const ChatMatchDuration = styled(MatchDuration)`
 export const InformText = styled(Text)`
   padding: 16px;
 `;
-
-export const ChatMessage = styled(Text)``;
 
 export const MessageContainer = styled(Flex)`
   flex-grow: 1;
