+++ src/pages/ChatRoomListPage/ChatRoomListPage.tsx
@@ -1,16 +1,16 @@
 import { Suspense } from 'react';
 
 import { Header } from '@components/Header';
+import { Flex } from '@components/shared/Flex';
 
 import { useChatRoomTabStore } from '@stores/chatRoomTab.store.ts';
 
-import { CHAT_ROOM_TAB_TITLE } from '@consts/chatRoomTabTitle.ts';
+import { CHAT_ROOM_TAB_TITLE } from '@consts/chat.ts';
 
 import { ChatRoomList } from './ChatRoomList.tsx';
 import {
   Main,
   MessagePageContainer,
-  TabBar,
   TabBarButton,
 } from './ChatRoomListPage.style.ts';
 import { SkeletonChatRoomList } from './SkeletonChatRoomList.tsx';
@@ -22,7 +22,7 @@ export const ChatRoomListPage = () => {
     <MessagePageContainer>
       <Header isLogo={false} title="채팅" isRightContainer={true} />
       <Main>
-        <TabBar>
+        <Flex>
           {Object.values(CHAT_ROOM_TAB_TITLE).map((tab) => (
             <TabBarButton
               onClick={() => setChatRoomTab(tab)}
@@ -32,7 +32,7 @@ export const ChatRoomListPage = () => {
               {tab}
             </TabBarButton>
           ))}
-        </TabBar>
+        </Flex>
         <Suspense fallback={<SkeletonChatRoomList />}>
           <ChatRoomList />
         </Suspense>
